Add files via upload
The game features colorful graphics and a simple, yet addictive gameplay that is easy to pick up, but hard to master. The obstacles come in the form of green pipes that are positioned at different heights, and the player must time their taps carefully to make the bird fly through the gaps between the pipes without colliding with them.

In addition to the basic gameplay, the Flutter version of Flappy Bird features several enhancements such as sound effects, high score tracking, and a game over screen that displays the player's score and allows them to restart the game.

Overall, the Flutter version of Flappy Bird is a fun and challenging game that is perfect for casual gamers looking for a quick and entertaining distraction.